requirements.txt
# backend
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.10.6
python-multipart
# libsql-experimental  # Temporarily disabled due to build issues with Python 3.9

# agentic
pandas==2.1.4
openai>=1.40.0
python-dotenv==1.0.0
httpx>=0.27.0,<0.28.0

# observability (optional - enable with OPIK_ENABLED=1)
opik>=1.0.0
